+++ tools/prune-trace/FESamplingPruner.hpp
@@ -1,6 +1,7 @@
 #ifndef __FESAMPLING_PRUNER_H__
 #define __FESAMPLING_PRUNER_H__
 
+#include <stdint.h>
 #include "Pruner.hpp"
 #include "util/CommandLine.hpp"
 
@@ -18,7 +19,7 @@ class FESamplingPruner : public Pruner {
 	fail::CommandLine::option_handle USE_KNOWN_RESULTS;
 	fail::CommandLine::option_handle NO_WEIGHTING;
 
-	unsigned m_samplesize;
+	uint64_t m_samplesize;
 	bool m_use_known_results, m_weighting;
 
 public:
